Add recipe and helper for deploying media
This adds a recipe and inc file for deploying test media. Media is
extracted from a tarball specified by SRC_URI, and known media types are
extracted and grouped in to packages, installed in /usr/share.

Audio filed can be installed by including test-multimedia-audio, video
files are in test-multimedia-video, etc.